custom redirect support
sign_in page now takes an rd= query parameter to use as the final redirect if
provided. This can include a full URL with scheme and hostname, to support an
oauth2_proxy serving several virtual hosts when being used as an NGINX auth
provider.

To use, follow the NGINX example in the readme but use this line for the error_page:

error_page 401 = /oauth2/sign_in?rd=$scheme://$server_name$request_uri;