Add X-Forwarded-User to /oauth2/auth response header
When acting as an NGINX auth proxy with auth_request, this lets you pass the
authenticated username to the proxied web service.

To do this, add the following two lines to your NGINX config next to your
auth_request line, and otherwise follow the example in the readme for nginx
auth_request.

    auth_request_set $user $upstream_http_x_session_user;
    proxy_set_header X-Forwarded-User $user;